[CHORE/#56] interceptor 의존성 분리
