+++ app/src/main/java/co/orange/ddanzi/di/module/InterceptorModule.kt
@@ -1,42 +1,24 @@
-package co.orange.ddanzi.di.interceptor
+package co.orange.ddanzi.di.module
 
 import co.orange.core.extension.isJsonArray
 import co.orange.core.extension.isJsonObject
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import co.orange.ddanzi.di.interceptor.AuthInterceptor
+import co.orange.ddanzi.di.interceptor.DeviceTokenInterceptor
+import co.orange.ddanzi.di.interceptor.RetrofitQualifier
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import kotlinx.serialization.json.Json
 import okhttp3.Interceptor
-import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.logging.HttpLoggingInterceptor
 import org.json.JSONArray
 import org.json.JSONObject
-import retrofit2.Converter
 import timber.log.Timber
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object InterceptorModule {
-    private const val APPLICATION_JSON = "application/json"
-
-    @Provides
-    @Singleton
-    fun provideJson(): Json =
-        Json {
-            ignoreUnknownKeys = true
-            prettyPrint = true
-        }
-
-    @Provides
-    @Singleton
-    fun provideJsonConverter(json: Json): Converter.Factory =
-        json.asConverterFactory(
-            APPLICATION_JSON.toMediaType(),
-        )
-
     @Provides
     @Singleton
     fun provideHttpLoggingInterceptor(): Interceptor =
